requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-crispy-forms==2.5
django-filter==25.2
djangorestframework==3.16.1
idna==3.11
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.2
